IMPORTANT UPDATE. ADDED A LIB FOR OAUTH TO GOOGLE. This way the backend handles the authentication.
This required flask-oauthlib

+++ SmartMirror.py
@@ -1,30 +1,87 @@
-from flask import Flask, request, render_template
+from flask import Flask, request, render_template, url_for, jsonify
 from db import DB
-from werkzeug.exceptions import BadRequest, NotFound
+from flask_oauthlib.client import  OAuth, session, redirect
+from werkzeug.exceptions import BadRequest
 from werkzeug.wrappers import BaseResponse as Response
 app = Flask(__name__, static_folder="static", static_url_path="/static",
             template_folder="templates")
 database = DB(app.root_path)
 
+app.config['GOOGLE_ID'] = "910617740457-56g7qqsturmt893oqd7ppornflq2lhs3.apps.googleusercontent.com"
+app.config['GOOGLE_SECRET'] = "mWLVEm4sjYa4BgyL-O5_fVXp"
+app.debug = True
+app.secret_key = 'development'
+oauth = OAuth(app)
 
-@app.route('/', methods=['GET'])
+google = oauth.remote_app(
+    'google',
+    consumer_key=app.config.get('GOOGLE_ID'),
+    consumer_secret=app.config.get('GOOGLE_SECRET'),
+    request_token_params={
+        'scope': 'email'
+    },
+    base_url='https://www.googleapis.com/oauth2/v1/',
+    request_token_url=None,
+    access_token_method='POST',
+    access_token_url='https://accounts.google.com/o/oauth2/token',
+    authorize_url='https://accounts.google.com/o/oauth2/auth',
+)
+
+
+@app.route('/')
 def index():
-    return render_template('index.html')
+    if 'google_token' in session:
+        me = google.get('userinfo')
+        return jsonify({"data": me.data})
+    return redirect(url_for('login'))
+
+
+@app.route('/login')
+def login():
+    return google.authorize(callback=url_for('authorized', _external=True)
+                            ,prompt='consent')
+
+
+@app.route('/logout')
+def logout():
+    session.pop('google_token', None)
+    return Response("User logged out", status=200)
+
+
+@app.route('/login/authorized')
+def authorized():
+    resp = google.authorized_response()
+    if resp is None:
+        return 'Access denied: reason=%s error=%s' % (
+            request.args['error_reason'],
+            request.args['error_description']
+        )
+    session['google_token'] = (resp['access_token'], '')
+    me = google.get('userinfo')
+    return jsonify({"data": me.data})
+
+
+@google.tokengetter
+def get_google_oauth_token():
+    return session.get('google_token')
+
 
 @app.route('/delete', methods=['GET'])
 def deleteUser():
-    email = request.args.get('email')
-    pin = request.args.get('pin')
-    if database.isUserRegistered(email):
-        try:
-            database.deleteUser(email, pin)
-        except BadRequest:
-            return Response("Wrong pin", status=403)
+    if 'google_token' in session:
+        email = request.args.get('email')
+        pin = request.args.get('pin')
+        if database.isUserRegistered(email):
+            try:
+                database.deleteUser(email, pin)
+            except BadRequest:
+                return Response("Wrong pin", status=403)
 
-        return Response("User deleted", status=202)
+            return Response("User deleted", status=202)
+        else:
+            return Response("User is not in the database", status=404)
     else:
-        return Response("User is not in the database", status=404)
-
+        return Response("NOT LOGGED IN", status=403)
 
 @app.route('/register', methods=['GET'])
 def enterRegistration():
